+++ ChangeLog.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - function `io/file-normalize-utf`
+- thread type support to the function `thread` to allow spawning daemon or user 
+  threads